+++ openbound/includes/Room.h
@@ -54,13 +54,13 @@ namespace SBURB
 
 		void SortDepths();
 
-		std::vector<std::shared_ptr<Action>> QueryActions(std::shared_ptr<Sprite> query, int x, int y);
+		std::vector<std::shared_ptr<Action>> QueryActions(std::shared_ptr<Sprite> query, float x, float y);
 		std::vector<std::shared_ptr<Action>> QueryActionsVisual(std::shared_ptr<Sprite> query, int x, int y);
 
 		bool IsInBounds(Sprite *sprite, int dx = 0, int dy = 0);
 		std::map<std::wstring, bool> IsInBoundsBatch(std::map<std::wstring, sf::Vector2f> queries, std::map<std::wstring, bool> *results);
 
-		std::shared_ptr<Sprite> Collides(Sprite *sprite, int dx = 0, int dy = 0);
+		std::shared_ptr<Sprite> Collides(Sprite *sprite, float dx = 0, float dy = 0);
 
 		std::wstring Serialize(std::wstring output);
